+++ src/ts-babel-generator.js
@@ -118,18 +118,18 @@ process.chdir(projectDir);
  * Install the necessary dependencies for typescript and babel.
  */
 if (!program.silent) console.info('Installing dependencies...');
-shell.exec('npm install --save-dev typescript @babel/core @babel/cli @babel/plugin-proposal-class-properties @babel/plugin-proposal-object-rest-spread @babel/preset-env @babel/preset-typescript @babel/plugin-proposal-numeric-separator', { silent: true });
+shell.exec('npm install --save-dev -s typescript @babel/core @babel/cli @babel/plugin-proposal-class-properties @babel/plugin-proposal-object-rest-spread @babel/preset-env @babel/preset-typescript @babel/plugin-proposal-numeric-separator', { silent: true });
 
 /**
  * Install any other dependencies that are needed.
  */
-if (program.webpack) shell.exec('npm install --save-dev webpack webpack-cli babel-loader');
-if (program.rollup) shell.exec('npm install --save-dev rollup rollup-plugin-babel@latest @rollup/plugin-node-resolve @rollup/plugin-commonjs');
+if (program.webpack) shell.exec('npm install --save-dev -s webpack webpack-cli babel-loader');
+if (program.rollup) shell.exec('npm install --save-dev -s rollup @rollup/plugin-babel @rollup/plugin-node-resolve @rollup/plugin-commonjs');
 
 /**
  * Set up the initial tsconfig.json config file.
  */
 if (!program.silent) console.info('Creating tsconfig.json file...');
-shell.exec('npm run tsconfig', { silent: true });
+shell.exec('tsc --init --declaration --allowSyntheticDefaultImports --target esnext --outDir lib', { silent: true });
 
 if (!program.silent) console.info('Finished!');